requirements.txt
# MCP Framework
mcp>=1.0.0

# HTTP Client
httpx>=0.25.0

# Environment Variables
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0
